Add a SIGKILL fallback if workers don't exit
When doing an upgrade, it's annoying if the old processes stick around
forever. While this clearly indicates a severe bug in the worker
code, it's nice if einhorn can be robust to this and not exacerbate the
situation.

I considered setting the default timeout to around 60s but decided not
to enable this behaviour by default. Sending SIGKILL is something you
should think pretty hard about, but it is (unfortunately) sometimes
necessary.